biding-validation

+++ src/main/java/com/auction/data/model/Auction.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.Date;
@@ -29,10 +30,10 @@ public class Auction {
     private String description;
 
     @Column(name = "start_price", nullable = false)
-    private BigInteger startPrice;
+    private BigDecimal startPrice;
 
     @Column(name = "buy_now_price", nullable = false)
-    private BigInteger buyNowPrice;
+    private BigDecimal buyNowPrice;
 
     @Column(name = "date_created")
     private LocalDateTime dateCreated;
